Improve type safety and error handling across the platform
Update type hints in several service files to use Optional and ensure correct float casting for confidence intervals.

+++ services/predictive_compliance_engine.py
@@ -541,8 +541,8 @@ class PredictiveComplianceEngine:
             
             # 95% confidence interval
             confidence_interval = (
-                max(0.0, forecast - 1.96 * std_error),
-                min(100.0, forecast + 1.96 * std_error)
+                max(0.0, float(forecast - 1.96 * std_error)),
+                min(100.0, float(forecast + 1.96 * std_error))
             )
         else:
             # Fallback confidence interval
